% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_lag.R
\name{custom_lag}
\alias{custom_lag}
\title{Lag a vector}
\usage{
custom_lag(x, k = 1)
}
\arguments{
\item{x}{Vector (of any class).}

\item{k}{Integer to specify the lag (defaults to 1).}
}
\value{
A lagged vector (not a time series as opposed to \code{\link[stats:lag]{stats::lag()}}).
The behaviour of this \code{lag()} function is closer to that of \code{dplyr::lag()}.
}
\description{
Lag a vector
}
